Update most demos for Godot 4.0.beta10
The HDR for 2D demo was renamed to Glow for 2D as autoexposure can
no longer be used in 2D in 4.0.

The SDF font demo was renamed to MSDF font to denote that multi-channel
distance field fonts are now used.

The OpenSimplexNoise demo was renamed to Noise Viewer as it now
uses FastNoiseLite.