Fix input not being detected when remap window is focused
If embedded windows are used, and RemapWizard/Mapping is focused, it
seems to be marking gamepad input as handled so it never reaches
remap_wizard.gd::_input in the parent node.
Disabling "handle input locally" in
RemapWizard/Mapping doesn't seem to have any effect.

Switching to native windows seems to fix the issue, but this will not
work for HTML5 platforms