Recreate tileset and tilemap; tweak bg
Create a tileset from scratch. The old one seemed completely broken, so
this seems like an improvement. Now you can jump around the level.

Change the mountain positions in background. I don't understand how this
is supposed to work, but now they show up in the background of the
level.