Setup parallax background
Move camera back to player.
Shrink window and setup camera limits so the camera moves around in the
world.

Copy parallax setup from 2d/platformer -> skeleton and use the same name
for the scene. That wasn't enough, so fiddle with parallax until it
fills the bg and I can see some movement. There's not much because the
camera doesn't have much space to move.