Upgrade 3d/voxel to Godot 4.4 UIDs for scenes and resources.
https://godotengine.org/article/uid-changes-coming-to-godot-4-4/

As of Godot 4.4, scenes and resources use UIDs to reference scripts and shaders.

Normally, this update is applied to a single scene or resource once you save it in Godot 4.4 for the first time. If you have a lot of scenes and/or resources, doing this manually may be time-consuming. This tool will update all of the project's scenes and resources at once.
Click "Restart & Upgrade" to restart the editor and update all of the scenes and resources in this project. Depending on the project size, it may take several minutes.
Note: Please make sure you have your project backed up before running the tool, to avoid the possibility of data loss. Additionally, make sure to commit all .uid files into version control (and do not add them to ignore-lists like.gitignore).